Big update here. The two podcasts are each displayed in their own tab, and each is downloaded from their respective bucket. The first time the user loads the app up, the titles need to be downloaded from scratch. Once this happens, they are saved to SharedPreferences, and displayed straight away the next time they open the app.
The bucket is now configured to send a message to an SQS queue when an
addition or deletion occurs in the bucket, meaning the podcasts need to
be updated. The parent Activity of the Fragments checks for these
messages on each startup, and if found, calls a method from within each
Fragment that re-downloads the data and updates the Adapter. The
messages are then deleted, meaning the next time the app starts up, the
podcasts are not downloaded again.